extract resource creation to getResourceFromMedia3p and add attribution
