+++ assets/src/edit-story/app/media/media3p/api/media3pApiProvider.js
@@ -22,7 +22,7 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import { createResource } from '../../utils';
+import getResourceFromMedia3p from '../../utils/getResourceFromMedia3p';
 import { listMedia as apiListMedia } from './apiFetcher';
 import Context from './context';
 
@@ -50,65 +50,6 @@ function Media3pApiProvider({ children }) {
     ]
       .filter(Boolean)
       .join(' ');
-  }
-
-  function getUrls(m) {
-    if (m.type.toLowerCase() === 'image') {
-      return Object.fromEntries(
-        new Map(
-          m.imageUrls.map((u) => [
-            u.imageName,
-            {
-              file: m.name,
-              source_url: u.url,
-              mime_type: u.mimeType,
-              width: u.width,
-              height: u.height,
-            },
-          ])
-        )
-      );
-    }
-    throw new Error('Invalid media type.');
-  }
-
-  function getFullAsset(m) {
-    if (m.type.toLowerCase() === 'image') {
-      if (!m.imageUrls.length) {
-        throw new Error('No imageUrls for media resource: ' + m);
-      }
-      // Reverse sort and get the largest width to determine the full asset.
-      return m.imageUrls.sort((el1, el2) => el2.width - el1.width)[0];
-    }
-    throw new Error('Invalid media type for media resource: ' + m);
-  }
-
-  /**
-   * Maps a media object returned from the API to a Story Editor resource.
-   *
-   * @param {Object} m The media object to map.
-   * @return {Object} The mapped resource.
-   */
-  function mapMediaToResource(m) {
-    const urls = getUrls(m);
-    const fullAsset = getFullAsset(m);
-    return createResource({
-      type: m.type.toLowerCase(),
-      mimeType: fullAsset.mimeType,
-      creationDate: m.createTime,
-      src: fullAsset.url,
-      width: fullAsset.width,
-      height: fullAsset.height,
-      poster: null, // TODO: Implement for videos.
-      posterId: null, // TODO: Implement for videos.
-      id: m.name,
-      length: null, // TODO: Implement for videos.
-      lengthFormatted: null, // TODO: Implement for videos.
-      title: m.description,
-      alt: null,
-      local: false, // TODO: How does this interact with the rest?
-      sizes: urls, // TODO: Map with expected keys for canvas.
-    });
   }
 
   /**
@@ -145,7 +86,7 @@ function Media3pApiProvider({ children }) {
       pageToken: pageToken,
     });
     return {
-      media: response.media.map(mapMediaToResource),
+      media: response.media.map(getResourceFromMedia3p),
       nextPageToken: response.nextPageToken,
     };
   }
