+++ assets/src/edit-story/app/media/utils/index.js
@@ -19,6 +19,7 @@ export { default as getFirstFrameOfVideo } from './getFirstFrameOfVideo';
 export { default as getResourceFromAttachment } from './getResourceFromAttachment';
 export { default as getResourceFromLocalFile } from './getResourceFromLocalFile';
 export { default as getResourceFromMediaPicker } from './getResourceFromMediaPicker';
+export { default as getResourceFromMedia3p } from './getResourceFromMedia3p';
 export { default as getTypeFromMime } from './getTypeFromMime';
 export { default as useUploadVideoFrame } from './useUploadVideoFrame';
 export { default as preloadImage } from './preloadImage';
